Corrections, added extractSolution Method, refactored readFile for imports
